requirements.txt
dvc
mlflow
opencv-python-headless 
numpy
Pillow
ultralytics             
fastapi
uvicorn
onnxruntime
streamlit